+++ crates/fluvio-cli/src/consume/mod.rs
@@ -37,11 +37,13 @@ use crossterm::{
 use crate::render::ProgressRenderer;
 use crate::{CliError, Result};
 use crate::common::FluvioExtensionMetadata;
+use crate::parse_isolation;
 use self::record_format::{
     format_text_record, format_binary_record, format_dynamic_record, format_raw_record,
     format_json, format_basic_table_record, format_fancy_table_record,
 };
 use handlebars::{self, Handlebars};
+use fluvio::dataplane::Isolation;
 
 const DEFAULT_TAIL: u32 = 10;
 const USER_TEMPLATE: &str = "user_template";
@@ -170,6 +172,12 @@ pub struct ConsumeOpt {
     /// Eg. fluvio consume topic-name --filter filter.wasm -e foo=bar -e key=value -e one=1
     #[clap(short = 'e', long= "extra-params", parse(try_from_str = parse_key_val), number_of_values = 1)]
     pub extra_params: Option<Vec<(String, String)>>,
+
+    /// Isolation level that consumer must respect.
+    /// Supported values: read_committed (ReadCommitted) - consume only committed records,
+    /// read_uncommitted (ReadUncommitted) - consume all records accepted by leader.
+    #[clap(long, parse(try_from_str = parse_isolation))]
+    pub isolation: Option<Isolation>,
 }
 
 fn parse_key_val(s: &str) -> Result<(String, String)> {
@@ -347,6 +355,10 @@ impl ConsumeOpt {
                     );
                 }
             }
+        }
+
+        if let Some(isolation) = self.isolation {
+            builder.isolation(isolation);
         }
 
         let consume_config = builder.build()?;