+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Set timestamp in Records while producing. ([#2288](https://github.com/infinyon/fluvio/issues/2288))
 * Support `ReadCommitted` isolation in SPU for Produce requests [#2336](https://github.com/infinyon/fluvio/pull/2336)
 * Improve error messages and add `--fix` option to `fluvio cluster check` to autofix recoverable errors ([#2308](https://github.com/infinyon/fluvio/issues/2308))
+* Producer must respect ReadCommitted isolation [#2302](https://github.com/infinyon/fluvio/issues/2302)
 
 ## Platform Version 0.9.24 - 2022-04-21
 * CLI: Migrate all fluvio crates to `comfy-table` from `prettytable-rs` ([#2285](https://github.com/infinyon/fluvio/issues/2263))